+++ contracts/mocks/introspection/facets/InterfaceDetectionFacetMock.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.17;
+pragma solidity 0.8.19;
 
 import {InterfaceDetectionStorage} from "./../../../introspection/libraries/InterfaceDetectionStorage.sol";
 import {InterfaceDetectionFacet} from "./../../../introspection/facets/InterfaceDetectionFacet.sol";