+++ CHANGELOG.md
@@ -10,10 +10,11 @@
 
 - Removed `ERC20SimpleMock`, `ERC20Mock` and `ERC20BurnableMock`.
 - Upgraded to `@openzeppelin/contracts@4.9.0`.
+- Upgraded to `solc@0.8.19`.
 
 ### Improvements
 
-- Update to latest dependencies.
+- Updated to latest dependencies.
 
 ## 1.1.1